Rename file models.ts to interfaces.ts
