+++ homework-8/src/interfaces.ts
